LPD-31288 Add StructureTreeContent component and calculate expanded keys outside that component to avoid unnecessary recalculations
StructureTreeContent uses hoverItem() so every time an item is hovered the expanded keys are calculated. Calculating expanded keys outside avoids unnecessary recalculations.
